Fixing issues with UI thread and data types

+++ WeatherApp/WeatherApp/Model/Weather.swift
@@ -17,7 +17,7 @@ struct Location: Codable {
     let name, region, country: String
     let lat, lon: Double
     let tzID: String
-    let localtimeEpoch: Int
+    let localtimeEpoch: Double
     let localtime: String
 
     enum CodingKeys: String, CodingKey {
@@ -30,20 +30,20 @@ struct Location: Codable {
 
 // MARK: - Current
 struct Current: Codable {
-    let lastUpdatedEpoch: Int
+    let lastUpdatedEpoch: Double
     let lastUpdated: String
-    let tempC: Int
+    let tempC: Double
     let tempF: Double
-    let isDay: Int
+    let isDay: Double
     let condition: Condition
     let windMph, windKph: Double
-    let windDegree: Int
+    let windDegree: Double
     let windDir: String
-    let pressureMB: Int
+    let pressureMB: Double
     let pressureIn: Double
-    let precipMm, precipIn, humidity, cloud: Int
+    let precipMm, precipIn, humidity, cloud: Double
     let feelslikeC, feelslikeF: Double
-    let visKM, visMiles, uv: Int
+    let visKM, visMiles, uv: Double
     let gustMph, gustKph: Double
 
     enum CodingKeys: String, CodingKey {
